refactor(admin/user_handlers): split crud.rs — create/update/delete → crud_write.rs (-263 LOC)
