+++ src/bin/email_api_dir/admin_ops/user_handlers/crud.rs
@@ -105,261 +105,6 @@ pub(crate) async fn api_admin_user_get(
     }
 }
 
-pub(crate) async fn api_admin_user_create(
-    req: HttpRequest,
-    body: web::Json<CreateAdminUserInput>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let actor = match admin_auth::require_admin(&req, &mongo, &mongo_db_name()).await {
-        Ok(u) => u,
-        Err(resp) => return resp,
-    };
-    let role = body.role.trim().to_ascii_lowercase();
-    if !["user", "admin", "support"].contains(&role.as_str()) {
-        return HttpResponse::BadRequest()
-            .json(serde_json::json!({ "message": "role must be user|admin|support" }));
-    }
-
-    let status = body
-        .status
-        .as_deref()
-        .unwrap_or("active")
-        .trim()
-        .to_ascii_lowercase();
-    if !["active", "restricted"].contains(&status.as_str()) {
-        return HttpResponse::BadRequest()
-            .json(serde_json::json!({ "message": "status must be active|restricted" }));
-    }
-
-    let id = body
-        .id
-        .clone()
-        .unwrap_or_else(|| format!("u_{}", Uuid::new_v4().simple()));
-    let now = now_iso();
-
-    let user = AdminUserRecord {
-        id: id.clone(),
-        email: body.email.trim().to_string(),
-        display_name: body.display_name.clone(),
-        role,
-        status,
-        two_factor_enabled: body.two_factor_enabled.unwrap_or(false),
-        last_login_at: None,
-        last_activity_at: Some(now.clone()),
-        sessions24h: 0,
-        actions7d: 0,
-        change_requests30d: 0,
-        recent_activity: vec![AdminUserActivity {
-            at: now.clone(),
-            label: "User created".to_string(),
-            kind: "admin_action".to_string(),
-        }],
-        created_at: now.clone(),
-        updated_at: now,
-        password_hash: None,
-        invite_token: None,
-        invite_expires_at: None,
-        invited_at: None,
-        notes: None,
-    };
-
-    let coll = mongo
-        .database(&mongo_db_name())
-        .collection::<AdminUserRecord>(ADMIN_USERS_COLL);
-
-    match coll.insert_one(&user).await {
-        Ok(_) => {
-            log_admin_action(
-                mongo.as_ref(),
-                &actor,
-                "user.create",
-                "admin_user",
-                &user.id,
-                None,
-                Some(serde_json::json!({ "email": user.email, "role": user.role, "status": user.status })),
-            )
-            .await;
-            HttpResponse::Created().json(serde_json::json!({ "user": user }))
-        }
-        Err(e) => {
-            eprintln!("api_admin_user_create error: {}", e);
-            HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "message": "Failed to create user" }))
-        }
-    }
-}
-
-pub(crate) async fn api_admin_user_patch(
-    req: HttpRequest,
-    path: web::Path<String>,
-    body: web::Json<UpdateAdminUserInput>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let actor = match admin_auth::require_admin(&req, &mongo, &mongo_db_name()).await {
-        Ok(u) => u,
-        Err(resp) => return resp,
-    };
-    let id = path.into_inner();
-    let coll = mongo
-        .database(&mongo_db_name())
-        .collection::<AdminUserRecord>(ADMIN_USERS_COLL);
-
-    let current = match coll.find_one(doc! { "id": &id }).await {
-        Ok(Some(v)) => v,
-        Ok(None) => {
-            return HttpResponse::NotFound()
-                .json(serde_json::json!({ "message": "User not found" }))
-        }
-        Err(e) => {
-            eprintln!("api_admin_user_patch read error: {}", e);
-            return HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "message": "Failed to load user" }));
-        }
-    };
-
-    let now = now_iso();
-    let mut updated = current.clone();
-
-    if let Some(role) = &body.role {
-        let role = role.trim().to_ascii_lowercase();
-        if !["user", "admin", "support"].contains(&role.as_str()) {
-            return HttpResponse::BadRequest()
-                .json(serde_json::json!({ "message": "role must be user|admin|support" }));
-        }
-        updated.role = role;
-    }
-
-    if let Some(status) = &body.status {
-        let status = status.trim().to_ascii_lowercase();
-        if !["active", "restricted"].contains(&status.as_str()) {
-            return HttpResponse::BadRequest()
-                .json(serde_json::json!({ "message": "status must be active|restricted" }));
-        }
-        updated.status = status;
-    }
-
-    if let Some(email) = &body.email {
-        let email = email.trim();
-        if email.is_empty() || !email.contains('@') {
-            return HttpResponse::BadRequest()
-                .json(serde_json::json!({ "message": "invalid email" }));
-        }
-        updated.email = email.to_string();
-    }
-    if let Some(dn) = &body.display_name {
-        let dn = dn.trim();
-        updated.display_name = if dn.is_empty() {
-            None
-        } else {
-            Some(dn.to_string())
-        };
-    }
-    if let Some(v) = body.two_factor_enabled {
-        updated.two_factor_enabled = v;
-    }
-    if let Some(notes) = &body.notes {
-        let notes = notes.trim();
-        if notes.len() > 1024 {
-            return HttpResponse::BadRequest()
-                .json(serde_json::json!({ "message": "notes too long (max 1024)" }));
-        }
-        updated.notes = if notes.is_empty() {
-            None
-        } else {
-            Some(notes.to_string())
-        };
-    }
-
-    updated.updated_at = now.clone();
-    updated.last_activity_at = Some(now.clone());
-    updated.actions7d += 1;
-    let note = body
-        .role
-        .as_ref()
-        .map(|r| format!("Role changed to {}", r))
-        .unwrap_or_else(|| "User updated".to_string());
-    let mut recent = updated.recent_activity;
-    recent.insert(
-        0,
-        AdminUserActivity {
-            at: now,
-            label: note,
-            kind: "role_change".to_string(),
-        },
-    );
-    recent.truncate(8);
-    updated.recent_activity = recent;
-
-    match coll
-        .replace_one(doc! { "id": &id }, &updated)
-        .upsert(false)
-        .await
-    {
-        Ok(_) => {
-            log_admin_action(
-                mongo.as_ref(),
-                &actor,
-                "user.patch",
-                "admin_user",
-                &id,
-                None,
-                Some(serde_json::json!({
-                    "role": updated.role,
-                    "status": updated.status,
-                    "email": updated.email,
-                    "displayName": updated.display_name,
-                    "twoFactorEnabled": updated.two_factor_enabled,
-                })),
-            )
-            .await;
-            HttpResponse::Ok().json(serde_json::json!({ "user": updated }))
-        }
-        Err(e) => {
-            eprintln!("api_admin_user_patch write error: {}", e);
-            HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "message": "Failed to update user" }))
-        }
-    }
-}
-
-pub(crate) async fn api_admin_user_delete(
-    req: HttpRequest,
-    path: web::Path<String>,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let actor = match admin_auth::require_admin(&req, &mongo, &mongo_db_name()).await {
-        Ok(u) => u,
-        Err(resp) => return resp,
-    };
-    let id = path.into_inner();
-    let coll = mongo
-        .database(&mongo_db_name())
-        .collection::<AdminUserRecord>(ADMIN_USERS_COLL);
-
-    match coll.delete_one(doc! { "id": &id }).await {
-        Ok(res) if res.deleted_count > 0 => {
-            log_admin_action(
-                mongo.as_ref(),
-                &actor,
-                "user.delete",
-                "admin_user",
-                &id,
-                None,
-                None,
-            )
-            .await;
-            HttpResponse::Ok().json(serde_json::json!({ "deleted": true, "id": id }))
-        }
-        Ok(_) => HttpResponse::NotFound()
-            .json(serde_json::json!({ "deleted": false, "message": "User not found" })),
-        Err(e) => {
-            eprintln!("api_admin_user_delete error: {}", e);
-            HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "message": "Failed to delete user" }))
-        }
-    }
-}
-
 /// PR1 (RBAC) — GET /api/admin/whoami
 ///
 /// Utilisé par le frontend pour connaître le rôle réel de l'utilisateur
