+++ src/bin/email_api_dir/admin_ops/user_handlers/mod.rs
@@ -1,7 +1,9 @@
 pub(crate) mod audit;
 pub(crate) mod crud;
+pub(crate) mod crud_write;
 pub(crate) mod lifecycle;
 
 pub(crate) use audit::*;
 pub(crate) use crud::*;
+pub(crate) use crud_write::*;
 pub(crate) use lifecycle::*;
